fix: syntax highlighting
