+++ package.json
@@ -28,9 +28,9 @@
     "clsx": "^2.0.0",
     "hast-util-is-element": "^1.1.0",
     "postcss": "^8.4.35",
-    "prism-react-renderer": "^2.3.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rehype-katex": "^7.0.0",
     "remark-math": "^6.0.0",
     "tailwindcss": "^3.4.1"
